Muhammad/server/requirements.txt
pandas==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
supabase==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.18
openpyxl==3.1.2
google-genai==0.8.5

